Add files via upload
This program performs a complete kinematic analysis of a fourbar mechanism with respect position analysis, velocity analysis and acceleration analysis, using the Freudenstein equations. 
The inputs are the four link dimensions: crank (a), coupler (b), rocker(c), and ground (d); the initial position of the crank (theta2) and the freeze frame angle (ffa). The outputs 
are the link angular positions for the coupler (theta3) and the rocker (theta4) and the positions in space for the link joints and coupler point; the angular velocities for the 
coupler (omega3) and the rocker (omega4) and the linear velocities of the link joints and coupler point; the angular accelerations for the coupler (alpha3) and the rocker
 (alpha4) and the linear accelerations of the link joints and coupler point.